requirements.txt
anthropic>=0.25.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.2.0
tzdata>=2024.1
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.128.0
